requirements.txt
langchain_openai
langchain_core
langchain-community
python-dotenv
streamlit
langserve
fastapi
uvicorn
sse_starlette